Add SetModule and fd_module defs
Summary:
This diff adds SetModule as a new AST def in the AAST. It does not implement parsing for it in the FFP, so nothing in the CST lowers to it just yet. But it allows us to represent setting module membership at the start of a file.

After this diff, we will add code in namespace_elaborate.rs to move namespaces to toplevel entities just like file attributes. This diff just adds the aast node and makes it compile.

Reviewed By: oulgen

Differential Revision: D35372038

fbshipit-source-id: 35d49da003c1b30eec90355231d0c2889586969b