Check that getInstanceKey can be called from PolicyShardedMemoize(LSB)-d functions
Summary:
We added an analogous check for Memoize(LSB) around the time
when coeffects (contexts and capabilities) feature was completed (Jan '21).

In the meantime, `__PolicyShardedMemoize(LSB)` attributes were added, which
do behave similarly with respect to memoization but, unlike `__Memoize(LSB)`,
do shard the cache by the implicit policy.  Therefore, they do not need
to check if for presence of `AccessGlobals` capability in the context.

Reviewed By: vassilmladenov

Differential Revision: D35834057

fbshipit-source-id: f57e3d9a617344b7c19cb00154e6968a64a1c6e1