Dynamic accesses affect both directions
Summary:
Previously, I implemented the affect of a dynamic access to a dict in forward dataflow direction thinking that a `dict` can be used as a `shape` up to the point of dynamic access at which point it can be converted to a `dict` with `toVec`.

However, this is not a great result and probably won't be actioned later on. So I changed the dynamic accesses to effect both forward and backward dataflow.

If we decide later that partial migration to shape is useful, we can revisit this.

Reviewed By: enetsee

Differential Revision: D33585215

fbshipit-source-id: 19a65e4dfc44cc7acfb60d9944b387a72726fdba