Ignore the enclosing expression type when hovering on type hints
Summary:
If the cursor is over a type hint, show information about the named type, not the enclosing expression.

This is particularly a problem on `is` expressions such as `$x is nonnull`. The expression has type `bool`, but if the user hovers over `nonnull`, we don't want to talk about `bool`.

Reviewed By: zhyty

Differential Revision: D35789410

fbshipit-source-id: 4e53b198d0df7c862397773d8fec05784e2fdba9