Remove uaInternal and replace with internal keyword everywhere
Summary:
This diff adds some lowering/decl support for internal on enums and typedefs, and in turn allows us to remove uaInternal and <<__Internal>> attribute entirely, now that the keyword is equivalent.

In doing this change I found a bug in typedefs in general, where internal newtype and internal type were treated equivalently even though they are not. I will preserve the broken behavior in this change (which is keyword only), and add a diff on top to fix that bug.

Reviewed By: chenmela

Differential Revision: D35824668

fbshipit-source-id: 35b66472c70252d0131d62c0efe2f7b99f7569c1