Make run.php more robust against workers dying
Summary:
Fail the test run and return a non-0 exit status if any test workers or
the output printing worker die due to a signal or exit with a non-0
status. Due to a bug in the existing implementation, workers that died
with a signal were not resulting in test run failure.

Reviewed By: ricklavoie, mofarrell

Differential Revision: D35851940

fbshipit-source-id: b77ee16f3fc936c09aaaa8d06b7f1e7b9e5cc0c9