Improve native Facts query performance for getAttributeType queries
Summary:
I was profiling a meerkat run and saw we were spending a bunch of time in `SQLiteAutoloadDBImpl::getTypesWithAttribute()`. Profiling the query in SQLite itself, I discovered we've been scanning `all_paths`.

Replacing this `WHERE EXISTS` monstrosity with a simpler query gives it an 80% speedup.

ashamed_bunny

Also, TIL that `WHERE EXISTS (SELECT * ...)` is a bit slower than `WHERE EXISTS (SELECT 1 ...)`. I expected them to be equivalent, but it's apparently the difference between `USING INDEX` and `USING COVERING INDEX` in SQLite's query planner.

Reviewed By: ricklavoie

Differential Revision: D35909018

fbshipit-source-id: eeb68167aa2f7c73a75f92a1855875f23f4858a3