remove anyfy
Summary: The `anyfy` function was used to propagate `Tany` types into the generic arguments. For example, when checking `Tany <: Traversable<#1>`, `Tany` would be added as a lower bound to `#1`. This suppresses subsequent type errors in a brittle and ad hoc way, by proliferating `Tany`s in the code. Instead, we will just consider `Tany` to be a subtype and super type of anything without further constraining any type arguments.

Reviewed By: CatherineGasnier

Differential Revision: D35776485

fbshipit-source-id: 9487f183ef42b20fe0686c19aa7ef09556875662