Opcode Translator macro outline
Summary:
Outline for how opcode translator will work.

Wanted to publish this first without all the ```IMM_XXX``` implementation in next diff to get any feedback, and to get it checked-in so it doesn't break as I continue working on it. (Can be tested by just building hhvm and ensuring macros are compatible with HackCUnit)

Differential Revision: D35841778

fbshipit-source-id: 02b1af117a0cbcecf90a4f4a8906733e9fdbd7d9