Make ext_curl set CURLOPT_POSTFIELDSIZE inside CurlResource::prepare
Summary: If we do not set `CURLOPT_POSTFIELDSIZE` to 0 when we have an empty POST, Curl will fail because it will try reading from a non-existent value. That gives a generic `read function returned funny value` error. Prior to this diff, we handled this case for regular Curl requests but did not do so for code using the curl_multi_ APIs. By moving the code into `prepare` we should handle it consistently for both.

Reviewed By: DavidSnider

Differential Revision: D35931456

fbshipit-source-id: 23fd3857b154fea542f436703bf187ac80302087