Prevent TAST check from returning escaping dicts
Summary:
Similar to parameters, we cannot make local deductions about dicts that are returned from functions (as they might be dynamically accessed later on).

Since we only want local results in the TAST mode, we add a dynamic constraint on the escaping dicts so that it gets filtered out when reporting results.

Reviewed By: enetsee

Differential Revision: D35895527

fbshipit-source-id: b46b521bfb5844b3f7252e66678fcd6ed6478914