Opcode Translator macro outline (#9075)
Summary:
Pull Request resolved: https://github.com/facebook/hhvm/pull/9075

Outline for how opcode translator will work.

Wanted to publish this first without all the ```IMM_XXX``` implementation in next diff to get any feedback, and to get it checked-in so it doesn't break as I continue working on it. (Can be tested by just building hhvm and ensuring macros are compatible with HackCUnit)

Reviewed By: edwinsmith

Differential Revision: D35841778

fbshipit-source-id: 8cce30a3406e10c97ccf6ac73afa26f532b4d6ad