Allow for 2 indirect call targets to be promoted to direct calls and inlined
Summary: This diff leverages the CallTargetProfile, which was only used to promote only the hottest call target. This diff also bumps the number of call targets tracked at each call site from 3 to 6, and reduces the threshold to trigger this optimization. We may want to generalize this to handle a configurable number of targets, but this already shows a measurable performance impact.

Reviewed By: paulbiss

Differential Revision: D35790014

fbshipit-source-id: de3a646b018ec249ad03e0e9452ce341a3dc393e