Only allow module membership statements before declarations in file
Summary:
This diff adds restrictions to where the statement `module foo;` can appear in a file.

Specifically, module membership statements must appear before any toplevel declarations in a file. This includes classes, typedefs, namespaces declarations, constants, and functions. They also must be unique in a file.

I've decided that namespaces should appear after module membership just because the following looks strange to me:
```
namespace Foo {
  module bar;
}
```

As an exception, we allow `new module foo {}` to appear before module membership. Probably the best rule for this is that `new module foo {}` can only appear in files without a module membership declarations, but it's very useful in tests to be able to define a module and then immediately use that definition, so we will allow
```
new module bar {}
module bar;
class Something {}
```
to parse for now.

Reviewed By: hgoldstein

Differential Revision: D35978613

fbshipit-source-id: 8c8085f1444e652ca89cacc7b7776580046ba04f