Removing SwitchableClass
Summary:
This class was a temporary measure to allow switching on enum class values.
Now that labels are available, it is safer to switch on a label instead.

This diff updates Hack only. As it will be released prior to HHVM, we can
forbid the usage of that class while HHVM is being deployed

Reviewed By: Wilfred

Differential Revision: D35995699

fbshipit-source-id: 6bc8b0416189e793a08393b457fc1ec43891412b