hackc rust binary (skeleton)
Summary:
This is an initial port of hh_single_compile_cpp's backbone, which uses bool
switches to dispatch to 9 possible subcommands.

Once the port is finished I'd like to use bonafide subcommands (an enum) since bool
options have 9 degrees of freedom and express 2^9 combinations, while an enum only
has one degree and 9 choices. Then it will be easy to merge in hh_compile's subcommands,
some of which are redundant.

Reviewed By: aorenste

Differential Revision: D35609894

fbshipit-source-id: d2e6f62633939d46ee67bf25d6a117e198aa5f66