Clean up make_return_type
Summary: First step in doing some clean up of `Typing_return.ml`. There are a bunch of functions that tend to get called as a set, when checking functions, methods, and lambdas. It would be nice to factor our this commonality - and will help some changes needed for Sound Dynamic.

Reviewed By: vsiles

Differential Revision: D36005306

fbshipit-source-id: 15fa8795159a83d551b064997547d7a16e5b58fe