fix fanout when adding a required XHP attribute
Summary:
There is a fanout bug that when we add an XHP attribute that is marked as required, we don't retypecheck corresponding XHP instances.

These instances are converted to constructor calls by the typechecker before being checked, so the fix in this diff leverages that.

The fix is implemented in Shallow_class_diff: when we detect that a required XHP attribute is added, we also mark the constructor as Modified.

Reviewed By: bobrenjc93

Differential Revision: D36097523

fbshipit-source-id: 50b089faaba11cb9f2fef71978e93a1950c5fbb4