Only kill hulk lite server if hulk activated
Summary: Now that we support incremental scenarios, I realized that we'll always kill hh_server, even if hulk wasn't activated. If the hulk lite flag is on, but we only do a local typecheck due to small fanout, we really shouldn't be killing the hh_server. This diff updates our behavior to only kill hh_server if hulk actually activated.

Reviewed By: nt591

Differential Revision: D36051740

fbshipit-source-id: 977ba705d8ffc948a095e8696b5b5a7c0b68830d