Add like to enforced return type when checking function body
Summary:
When a return type is enforced, we can check the body of the function with a more liberal like-type. This permits some flexibility with use of like types:for example, it lets us write `return foo()[0];` where `foo` has a return type of `vec<~int>`.

We deliberately support this behaviour only for `<<__SupportDynamicType>>` functions, as in a future "fully sound" world we may wish to remove automatic enforcement for functions *not* marked `<<__SupportDynamicType>>` without affecting behaviour.

Differential Revision: D36095174

fbshipit-source-id: ceb2b8717fe3c6c0e249841beff942973542ccfd