add facts --nohash subcommand
Summary:
Somtimes (e.g. for RE), we do not want the source hash to be part of the Facts output;
doing this unnecessarily makes the output change when the input has boring changes
(e.g. just comments). Avoiding poisoning the output is crucial for short-circuiting.

To accomplish this:
* add a `hackc facts` subcommand which will eventually subsume
the `--extract-facts-from-decls` flag command.
* add a `hackc facts --nohash` option to the subcommand that serializes
Facts as json without the `sha1sum` field.
* And reorganize Opts so the `--input-file-list` option and `files` positional
argument can be reused by subcommands.

Reviewed By: aorenste

Differential Revision: D36144142

fbshipit-source-id: 75c4fba5ffec6ff71082adac7d18ec42cdfbeb28