Build HHVM packages both as an impure-derivation and an ordinary derivation (#9084)
Summary:
The impure derivation is built in a Nix sandbox, which ensures the build script did not accidentally reference any non-Nix dependencies.

The default package is not an impure derivation so it is cacheable, but it is built without a Nix sandbox because it requires internet access.

Pull Request resolved: https://github.com/facebook/hhvm/pull/9084

Differential Revision: D36203430

Pulled By: Atry

fbshipit-source-id: 0ddc6bceb6cf4d44398563117b1984f64811c859