Include memory effects when calculating defining instruction equivalency
Summary:
For rematerialization, vasm-graph-color first tries to find a
single defining instruction for the Vreg. If the Vreg has multiple
definitions (which can happen if phis are involved), it checks if all
of them are "equivalent", and picks an arbitrary one. This should be
okay, since we've checked they're all equivalent. If the
rematerialization involves a memory read, we use this canonical
defining instruction to calculate memory effects. The memory effects
are used to see if there's been any intervening stores between the
original load and the rematerialization point.

However, we were not consulting memory effects when seeing if multiple
defining instructions are equivalent. It is possible for two memory
reading vasm instructions to be otherwise identical, but represent
different memory effects. For example, loads of a property may happen
to be at the same offset from an object, but represent different
properties. Since we choose an instruction arbitrarily, we'll only use
one of the memory effects arbitrarily to see if the reload is
safe. This means we can miss legitimate writes which should invalidate
the reload.

Fix by including memory effects in the equivalency analysis. If the
instructions represent memory loads, check their origin. If their
origins are different, calculate the memory effects from the origins
and see if they are equivalent. If they are not, act as if the
instructions are not equivalent.

We can technically do better than this, and calculate the union of the
memory effects along the different branches, but this would involve
bigger changes to the rematerialization logic.

Reviewed By: jano

Differential Revision: D36230728

fbshipit-source-id: cafb728ef9879b912c8578e9937884bac5594427