Replace instrJumpOffsets() with instrJumpTargets()
Summary:
instrJumpOffsets() returned an offset relative to the opcode. While this
matches how offsets are currently encoded in the bytecode stream, it is
generally not very useful.

Replace all its remaining users with instrJumpTargets(), which returns offsets
relative to the function entry. Merge instrJumpOffsets() implementation into
instrJumpTargets().

Allows introduction of the new Entry opcode that jumps to offset zero.

Reviewed By: ricklavoie

Differential Revision: D36230782

fbshipit-source-id: cd6603f3d29c64132f67de41a79855c7c23be9dc