`hackfmt` does not respect `@partially-generated` tag if there are no manual sections
Summary:
`hackfmt` special cases files with the `partially-generated` header to ensure that formatting is applied to manual sections only. However, the test that is used does not correctly handle the case of a partially generated file without any manual sections (which should be equivalent to a fully generated file) and applies formatting to the whole file instead.

This diff corrects this by treating partially generated without manual sections as though it were fully generated.

Differential Revision: D35465987

fbshipit-source-id: 31fa71fbc554069f736e456fcd20c20faf56d3d0