merge emitter crates
Summary:
This is a step in pursuit of a less granular crate structure for hackc,
but still granular enough. Starting with emit_unit which is the primary
API for the Emitter, merge crates that are only used by emit_unit,
and whose purpose seemed tightly coupled to emitting units.

There may be more we could merge, like label_rewriter, but it felt
like a good stopping point. We might want to preserve some kind of
division between "business logic" (aast -> hhbc) and generic transformations
like label_rewriter. E.g. if this were a more formal IR we'd want a crate
for passes, or possibly even individual crates for heavy duty passes.

Reviewed By: shayne-fletcher

Differential Revision: D36202101

fbshipit-source-id: 60d8cf86991e8b2d61e48499b91b79d455e4ccdc