Stop creating manual mock hashes in SymbolMapTest
Summary:
I'm changing around some code in `SymbolMap` and finding that `SymbolMapTest` fails due to internal implementation changes.

In particular, my changes assume that if the SHA1 hash is the same, then the facts must also be the same. But since this test sometimes stores different facts with the same hash of `"aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"`, this assumption is broken.

We can make this assumption hold automatically, and we can avoid the need to constantly write `"aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"`, by automatically creating SHA1 hashes from the facts themselves.

Reviewed By: edwinsmith

Differential Revision: D36257837

fbshipit-source-id: 5065e8bfb344672d15c7db4d4a46f938a786b03d