Remove a flaky test from code synthesis suite
Summary:
This particular test is flaky on master because there are multiple pieces of code that can satisfy its constraints for code synthesis. In particular, any combination of S0-S8 interfaces will not affect the final result of dependency graphs as long as theres the correct number and average width.
Delete it for now; we have good coverage with end to end tests here that cover the same thing in a less flaky way.

Reviewed By: bobrenjc93

Differential Revision: D36324071

fbshipit-source-id: d84a2a30ca61b42756af5b0e663c1a371c1577c3