Factor out make_folded_decl_provider
Summary:
There's some overlapping code in various places where we attempt to do the following:
- Create a shallow decl cache (with various configs)
- Populate the shallow decl cache given a list of filenames
- Create a folded decl provider given a shallow decl cache

This diff does the latter two refactoring.

A couple of notes. When populating the shallow decl cache given a list of files, I use parallel iterators with a progress bar. This is different from some codepaths originally. I figure this is okay because, populating the shallow decl cache "should" be order independent so it can be done in parallel and because the indicatif library seems to suggest that if it detects that it is not printing the progress bar to console via stderr, it won't print the progress bar at all. So hopefully no one is affected in workflows that are not command line.

Reviewed By: shayne-fletcher

Differential Revision: D35406417

fbshipit-source-id: a8e76062320479de7305000be3bb55b8dbb840af