Factor out make_shallow_decl_cache
Summary:
There's some overlapping code in various places where we attempt to do the following:

1) Create a shallow decl cache (with various configs)
2) Populate the shallow decl cache given a list of filenames
3) Create a folded decl provider given a shallow decl cache

This diff currently just factors out the "create a shallow decl cache part". The next diff does the latter two refactoring.

This diff and the next should try and create a more common workflow for these three steps that is hopefully general enough to encompass future efforts. If we prefer not to do it this way, also happy to repeat code, just didn't want to if we decided that something should change about any of these steps.

Reviewed By: shayne-fletcher

Differential Revision: D35406418

fbshipit-source-id: ddbe14d5d89a8e51933b470a00685f9dc1e9f74f