new cargo_vendor
Summary:
The old version of cargo_vendor.sh used `cargo vendor` to pull in the crates locally and then push that tarball to manifold.

This was problematic because the versions of the crates that were pulled in almost never matched the versions that we were using in fbcode.

In addition it required an internet connection to fetch the crates from crates.io.

This rewrites that script to copy the crates directly from fbcode using the same versions as we're testing in fbcode.

Unfortunately a when configured this way `cargo vendor` has it's own bug workaround that causes it to delete our vendor - so I needed to copy a couple of smalll files from `cargo vendor` locally to tweak them.

Differential Revision: D36213922

fbshipit-source-id: f7dc7d59b917ce2561219282e8b6f87eea40571f