Add ffi function to shallow decl and fold a partition of a directory
Summary:
Previous, I attempted to fold www without ooming by writing the computed folded decls to disk to avoid storing them all in memory. This worked on my devserver... But it doesn't work on sandcastle. Even though the requests don't always OOM - the sandcastle workers die.

Perhaps this is because the machines that have enough RAM to hold the shallow decls in memory and www don't have any hard drive space and instead are all flash. At any rate, after fumbling with that approach for a while, instead, the idea is to partition the classes in www and fold over them in segments - hoping that this goes under the OOM bar on these XLARGE sandcastle machines.

This diff adds the rust logic that does this. Instead of writing stuff to disk (which was very slow), just keep everything in memory and pass it back from rust to ocaml.

Reviewed By: shayne-fletcher

Differential Revision: D36335637

fbshipit-source-id: 2e5a56db53e6edb85bc8e27012e1320f1e3c6600