Add ocaml logic
Summary:
OCaml part of the comparator.

The logic:

- Calls into rust to collect all of the shallow decls of the directory, then folds over a partition of the classes found. Rust passes back a map of the folded classes
- In OCaml, collect all of the shallow classes. For each folded class in rust, compute the folded class in OCaml and then compare the two. If the two are different (according to rust), then write the decls to a file in the output dir.

The OCaml binary takes the following flag args:
```
      ( "--output-dir",
        Arg.String (fun s -> output_dir := Some s),
        " Path to write decl differences to" );
      ( "--partition-index",
        Arg.Int (fun i -> partition_index := i),
        " What section of folded decls we compare" );
      ( "--num-partitions",
        Arg.Int (fun i -> num_partitions := i),
        " How many ranges there should be" );
      ( "--print-limit",
        Arg.Int (fun i -> print_limit := i),
        " How many diffs should be written per OCaml process" );
```
And an additional required non flag argument to the directory that should be folded over.

The print limit argument is so that if every decl in the repo is different, then we do not kill the sandcastle machines since they don't have hard drive space and instead write everything to memory.

The partition flag arguments is so that we don't oom or kill sandcastle workers by attempting to fold over the entire repo.

Reviewed By: shayne-fletcher

Differential Revision: D36335638

fbshipit-source-id: 0daf9b2aae57f501cea830ac11c79b8600f33731