Use func families more aggressively even when we don't know the class in HHBBC
Summary:
Right now func families are only used when we can actually resolve the
class type to something specific. We can go one step further. We can
build a second set of func families, representing methods with a
particular name across all classes. Thus, even we if don't know the
class, we can still use that set of func families. Naturally this set
won't be as precise, but its better than nothing. This removes the
need for the few adhoc tables we had constructed which were trying to
do a similar thing.

There's one important difference between these new func families and
the old ones. With the old ones, if you resolved to one with a single
method, it was guaranteed to exist (in fact exactFunc() would return a
func). With the one ones, this isn't the case. Just because you're
calling a method named "foo", and there's only one method in the
program named "foo", doesn't guarantee that's what you're calling. The
call might fail. To distinguish this, add a new variant to res::Func,
named MethodOrMissing. It behaves like a MethTableEntry, except
exactFunc() won't return a func for it.

While testing this, I realized that we weren't adding methods from
flattened traits in the index's methods table, which is bad. Do
that. The methods table in the index isn't actually used after
building the index, so clear it out beforehand (this saves some memory
for the analysis phase).

Reviewed By: jano

Differential Revision: D36271855

fbshipit-source-id: c818047469632efb292460faeb1e793bb635308b