Rewrite ternary operator to "if then else" to eliminate call to Variant copy constructor
Summary:
Rewrite C ternary operator to use "if then else" to prevent Clang from invoking copy constructor  when the else path is taken.

gcc is able to handle the ternary without copy constructor while Clang is not. We are looking to improve Clang. For the time being fix these instances,

Reviewed By: ottoni

Differential Revision: D36182509

fbshipit-source-id: 88a23ee16aae453c6ecba5bc506952491fd7c49a