Use HHIs in `hphp/runtime/ext/` to stub out constants for typechecking
Summary: While attempting to typecheck systemlib, we encouter cases where we need to be aware of a constant's existence in order to correctly typecheck function / class definitions. Previously we'd just copy-paste the constant definition into PHP and leave the constant with two definitions: one in C++ and one in PHP. This allows to avoid having two "real" definitions: we include an HHI while typechecking that is not included while actually building systemlib.

Reviewed By: edwinsmith

Differential Revision: D36297678

fbshipit-source-id: f442bd1aaa927cdd7ec11baaf20085faced4fde9