Revise Rust_provider_backend.Naming FFIs to work with Db_path_provider
Summary: Rather than having each Naming FFI accept a db_path argument, as Naming_heap does, make the FFI stateful, with a function to set the current DB path (with the intent that we'll invoke it in the Rust_provider_backend implementation of Db_path_provider). This ensures that when we use NamingProvider internally on the Rust side, we'll use the correct database (since a db_path is _not_ threaded through the Rust naming_provider::NamingProvider trait methods, and I think it's best not to).

Reviewed By: shayne-fletcher

Differential Revision: D36302665

fbshipit-source-id: 91a2dbcd53d703fd8e037cd9f46fedcc9fc8b8cb