C++ decl-provider API cleanup
Summary:
The C++ struct DeclProviderResult was handwritten in two places (hhvm
and hh_single_compile_cpp), and must be kept in sync with the rust
enum ExternDeclProviderResult. Move the C++ decl to a header file,
and provide constructors that know how to fish into a DeclResult.

The philosophy behind this change is for hackc to provide a [relatively] clean
C++ API to its C++ clients, with a goal of hiding some of the uglier aspects
of C++/Rust bridging behind that api. There is more we can do after this.

CXX understands Vec<u8>, so use that instead of Bytes in DeclResult and a few
related APIs. This removes the last use of ffi::Bytes in the cxx-based compiler_ffi,
and more clearly represents DeclResult's ownership of the serialized bytes.
hackc_verify_deserialization() is now safe.

Merged FileAttributes and Bump into Decls, so that by-ref types and the arena
that owns them are grouped together in a single Rust type that is opaque to C++,
following the same pattern we have used elsewhere.

Reviewed By: shayne-fletcher

Differential Revision: D36423632

fbshipit-source-id: d36a32c80df87bb1caf0953ac1abb9a3491febf4