Parse `Type with { ... }` syntax as unstable feature
Summary:
Implements the provisional syntax for parsing so-called *type refinements*:

  <Type> with { <RefinementMembers> }

where <RefinementMembers> contain members in the syntactic form:

  type <name> <TypeConstraints>
  type <name> = <Type>
  ctx <name> <ContextConstraints>
  ctx <name> = <ContextList>

each member being separated by a `;`.  The trailing `;` is optional to
reduce verbosity for the most common case of a type refinement with 1 member
that doesn't require line-splitting, e.g.:

  function takes_int_box(Box with { type T = int } $b): void

but `;` is encouraged in a multi-line refinement such as:

  Box with {
    type T1 as MyBaseClass;
    type T2 = int;
    ctx C super [defaults] as [read_globals];
  }

Don't split in Hack format if the entire type + refinement fits in 1 line, e.g.

  function box_of_box_of_int(
  ): Box with { type T as Box with { type T = int } };

is preferred to

  function box_of_box_of_int(): Box with {
    type T as Box with { type T = int }
  };

because the latter could be misread as a function returning a `Box` that
has a gibberish (misparsed) body, especially in abstract method declarations.
The keyword `with` is required for the same reason besides simpler parsing.

Gate this feature in the parses using the `EnableUnstableFeature` framework.

Reviewed By: mpu

Differential Revision: D36300132

fbshipit-source-id: 30b66ca00cc21d157588e2b1e247ce3f35c02fa4