Move responbility of updating coerced value from fallback() to verifyTypeImpl()
Summary:
Use the recently introduced setVal() callback to update the value after
possible coercion when handling the FallbackCoerce case.

This also stops usage of VerifyProp when we can possibly coerce. We now use the
VerifyPropCoerce opcode. These are currently identical C++ calls, but the
non-coerce version can be optimized in the future.

Reviewed By: oulgen

Differential Revision: D36463284

fbshipit-source-id: a6b6044fceb611c4884baf3b675bbb444d934bc1