Move hackrs::cache to separate crate; rename to store
Summary: Move these traits and helpers to a separate crate so that they may be used in crates like `file_provider`. Rename `Cache` to `Store` (and likewise for all related names), since "cache" is a bit of a misnomer (for instance, the naming SQLite DB is clearly a datastore, but it's tougher to argue that it is a *cache*).

Reviewed By: shayne-fletcher

Differential Revision: D36465269

fbshipit-source-id: 478448055364b2f88baade45f4111ce45f427133