Refactor ChangesCache to use fallback Cache
Summary:
Instead of using the `Lookup` type to communicate whether an element was present, deleted, or unknown to the changes stack, accept a fallback `Cache` representing what lies underneath the temporary-changes stack (i.e., in hh_server, sharedmem and the naming SQLite database). This allows us to perform a lookup from the fallback in what was previously the `Lookup::Unknown` case, and we can return `Option<V>` instead of `Lookup<V>`.

This requires us to express our database queries in terms of the `Cache` trait, so the `TypeDb`, `FunDb`, (etc) wrapper structs are added to do so.

Reviewed By: shayne-fletcher

Differential Revision: D36465268

fbshipit-source-id: b35c450714795f71225d2f9b6b0efa9b2418a3f2