Add DeltaStore
Summary: Add a store helper which maintains a mutable delta in front of a readonly data source. Will be used to model mutable hh_server sharedmem in front of the naming table DB. Differs from ChangesStore in that only one delta is maintained instead of a stack. I also intend to add write-through behavior to ChangesStore when the stack is empty, but DeltaStore will never write to the readonly fallback.

Reviewed By: shayne-fletcher

Differential Revision: D36465271

fbshipit-source-id: 801cde0daa7a174dc0915932bee513497d054e86