core
Summary:
This is the initial version of the IR core.  The IR core contains all the basic structure of the CFG but has no auxiliary helpers (such as printing) or passes (such as rpo sorting).

The IR is intended to eventually serve as a intermediate form between AST and bytecode which is lower-level than the AST but easier to manipulate than the bytecode.

Reviewed By: shayne-fletcher

Differential Revision: D34371878

fbshipit-source-id: de6b78e5ef395f5743834941541c150438fd9092