Create and use ClassHasReifiedGenerics IR instruction
Summary:
IR instruction to determine if a class has reified generics; this will be used in D35663015 so the JIT checks can match the checks done in the interpreter.

This also moves the checks that were being done in `newInstanceReified()` out to place in the interpreter where `newInstanceReified()` was being called, and to generates the `classHasReifiedGenerics` instruction before the places that are generating an `allocObjReified` instruction (since `allocObjReified` generates a native call to `newInstanceReified()`).

Reviewed By: oulgen

Differential Revision: D36181834

fbshipit-source-id: 483f176c47577d4f78537342fee114fd8923dfda