Remove hackc_unit_to_string_cpp_ffi
Summary:
At first I converted hh_single_compile_cpp --compile_and_print_unit to
a simple gtest. Then I realized this is just a gtest for an FFI wrapper for
bytecode_printer. Why do we need that?

The only real use of it was for HackCTranslator round-trip testing, which
compares hackc->unit->bytecode_printer with hackc->translator->disasm.
But the A side of this comparison is identical to an existing FFI which is
already called from the same place 15 lines earlier.

Longer term I don't think we need standalone FFI for just bytecode_printer,
so this diff refactors it out of existence, then removes the CLI option.

Reviewed By: kavoor

Differential Revision: D36412698

fbshipit-source-id: 1c47c77d5b528ea179181c41617e137bac94af24