Typecheck `ext_std_classobj`
Summary:
Let's get `ext_std_classobj` well typed: we mainly want this for `get_class`, but this also uncovers some other issues.
 ---
Our logic for determining the expected function signature for a `__Native` function was slightly off. HHVM cannot, for it's native interface, represent nullable types: "nullability" is as good as `mixed`. Unfortunately, this logic would fall over on "nullable collections" (e.g.: nullable types with a `<` in their name). It would prioritize the collection part, assume the output could never be null, and promptly crash when it was. The only callsite here is in `hphp/runtime/vm/as.cpp`, in `type_constraint_to_data_type`, so I ended up just in-lining and simplifying the logic: we no longer check if something is a "collection." Assuming nobody is doing something like:
```
function foobar(); arraykey<Bing>;
```
The logic seems equivalent here.
 ---
`object` is a PHP type that used to mean "all classes;" it's since been killed in Hack given it was horribly unsound (you could call any method on it / access any prop, and you'd get back `TAny`). I've added a stub type `object` that aliases to `dynamic`. This is good enough for the typechecker, and close to morally correct, and allows HHVM to retain it's current behavior ... which is that unknown types in native functions are treated as if they're object names.

Reviewed By: aorenste

Differential Revision: D36355501

fbshipit-source-id: 157c67e1657efd921f8ee48b63ef60b4a0feb226