Add TargetProfile to coeffect fun param decision
Summary:
Looking at tcprints, I have noticed that due to sheer number of possible types `[ctx $x]` can take, we end up emitting a lot of code. Our main/cold relocation optimizations help but we still end up with a ton of branches. In most cases, we can profile exactly the branch we need and when we can't find the exact one, we can always sort them in order of likelihood.

This target profile massively decreases the code size of coeffect fun param.

Reviewed By: ricklavoie

Differential Revision: D36076820

fbshipit-source-id: 64822c78e72c3206c30d15079ff41b7a2f055666