Add slot based optimizations to loading context constants
Summary:
When we try to lookup a context constant, if we have the exact class, we can just lookup and load the context constant ahead of time. However, if we only know the base type of the class, then we were not doing any optimizations.

Let's figure out the slot of the context constant and jit loading from that slot for the derived class. This avoids having to call a C++ function to lookup the context constant.

Reviewed By: mofarrell

Differential Revision: D36073252

fbshipit-source-id: e0b8abf2dcd4191c8dd34f985db93da2a71ea9eb