Support `super` on type consts as unstable feature
Summary:
Add ability to specify lower bounds via `super` on type constants,
analogously to what is currently allow for context constants. E.g.:
```
abstract class A {
  abstract const type T super int = arraykey;
  abstract const type Tlsb as A super this;
}
class B extends A {
  const type Tlsb as B super this = this;
}
```

The lowerer partitions `as`/`super` constraints into lower and upper sets,
computing union type of the former and intersection type for the latter. E.g.
```
abstract class C {
  // NOTE: lowers into `T super (int | A::T)`;
  abstract const type T super int super A::T;
}
```

*NOTE*: **Typechecking comes "for free"** -- no changes required!

For backward compatibility, ban this in the typechecker (unless opted-in from tests using --allow-unstable-features),
but allow them in the HackC/HHVM parser (so we don't slow the rollout later)
because bounds on type constants are invisible to the HHVM anyway.

Reviewed By: vassilmladenov

Differential Revision: D35976902

fbshipit-source-id: 771e939e31316e66b1cbc5bb0acb0ea80fe88b31